TRy to speedup test (and skip Python 3.15; orjson).
Trying to move some test to latest only (networkinterface is 70s on
macos!!).

orjson does not install on 3.15, so bin pre-release to 3.14,
and increase hearbeat rate in the hope of speeding up restart and dead
detection tests.

+++ tests/problemkernel.py
@@ -34,7 +34,7 @@ class ProblemTestApp(IPKernelApp):
 
 if __name__ == "__main__":
     # make startup artificially slow,
-    # so that we exercise client logic for slow-starting kernels
-    startup_delay = int(os.environ.get("STARTUP_DELAY", "2"))
+    # so that we exercise client logic for slow-starting kernels.
+    startup_delay = float(os.environ.get("STARTUP_DELAY", "2"))
     time.sleep(startup_delay)
     ProblemTestApp.launch_instance()
